[Gaarahan] feature: add valid for add rs-event and complete delete rs-event

+++ src/main/java/com/thoughtworks/rslist/api/RsController.java
@@ -67,6 +67,10 @@ public class RsController {
   @Transactional
   @DeleteMapping("/rs/{id}")
   public void deleteRsEventById(@PathVariable int id) {
-    this.rsEventRepository.deleteById(id);
+    if (this.rsEventRepository.findById(id).isPresent()) {
+      this.rsEventRepository.deleteById(id);
+    } else {
+      throw new InvalidUserIdException();
+    }
   }
 }
